backend/requirements.txt
# =============================================================================
# Backend Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# --- Web Framework ---
fastapi==0.115.0          # Modern async web framework
uvicorn[standard]==0.30.6 # ASGI server to run FastAPI

# --- Database ---
sqlalchemy[asyncio]==2.0.35  # ORM with async support
asyncpg==0.30.0              # Async PostgreSQL driver
alembic==1.13.3              # Database migration tool
greenlet==3.1.1              # Required by SQLAlchemy async

# --- Authentication ---
passlib[bcrypt]==1.7.4    # Password hashing
python-jose[cryptography]==3.3.0  # JWT token handling
python-multipart==0.0.12  # Form data parsing (login forms)
authlib==1.3.2            # Google OAuth integration
httpx==0.27.2             # Async HTTP client (OAuth flows)
itsdangerous==2.2.0       # URL-safe token signing (magic links)

# --- Notifications ---
aiosmtplib==3.0.2         # Async SMTP for email
twilio==9.3.0             # SMS via Twilio
jinja2==3.1.4             # Email template rendering

# --- Task Scheduling ---
apscheduler==3.10.4       # Background job scheduler
celery[redis]==5.4.0      # Distributed task queue (production)
redis==5.1.1              # Redis client for Celery backend

# --- Utilities ---
pydantic[email-validator]==2.9.2  # Data validation
pydantic-settings==2.5.2  # Settings from env vars
python-dotenv==1.0.1       # Load .env files

# --- Development ---
pytest==8.3.3             # Testing framework
pytest-asyncio==0.24.0    # Async test support
httpx==0.27.2             # Test client
